Sidebar toggle event listener to modify wrapper z-index
Set default #wrapper z-index to -1 instead of 0, so that is behind the body by default, to allow interaction with the affordances (images and download links).

Added an eventListener in the javascript to modify the wrapper z-index from -1 to 0 and vice-versa when sidebarToggle is clicked.

Made it so that the #wrapper transition from z-index 0 to -1 happens only after the css transition of side #sidebar-wrapper margin-left is done, using setTimeout.

Added comments explaining these changes.